fix(docgen): create snippets only where supposed to.
