+++ docgen/snippets/updateAllSnippets.ts
@@ -1,21 +1,10 @@
-import path from "path";
-import { fileURLToPath } from "url";
-import getComponentDirs from "../helpers/getComponentDirs.js";
+import getSnippetPagePaths from "./getSnippetPagePaths.js";
 import updateSnippetsForPage from "./updateSnippetsForPage.js";
 
-const dirname = path.dirname(fileURLToPath(import.meta.url));
-const rootDir = path.resolve(dirname, "../../src/routes/components");
-
 export default async function updateAllSnippets() {
-  const componentsToIgnore = ["layout"];
+  const snippetPagePaths = await getSnippetPagePaths();
 
-  const componentDirs = (await getComponentDirs(rootDir)).filter(
-    (dir) => !componentsToIgnore.includes(dir)
-  );
-
-  for (const dir of componentDirs) {
-    const dirPath = path.resolve(rootDir, dir);
-    const pageFilePath = path.resolve(dirPath, "+page.svelte");
-    await updateSnippetsForPage(pageFilePath);
+  for (const snippetPagePath of snippetPagePaths) {
+    await updateSnippetsForPage(snippetPagePath);
   }
 }
